# Speedup Trilogy#escape by 3 to 5x